refactoring: 프로젝트 내에서 GlobalException 발생 시, GlobalExceptionHandler 에서 GlobalException 의 에러 타입에 따라 처리하도록 구현
