+++ src/main/java/com/example/springboard/common/handler/GlobalExceptionHandler.java
@@ -2,10 +2,8 @@ package com.example.springboard.common.handler;
 
 import com.example.springboard.common.ErrorResponse;
 import com.example.springboard.common.exception.GlobalException;
-import com.example.springboard.util.enums.ErrorCode;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
@@ -17,8 +15,14 @@ public class GlobalExceptionHandler {
 
     @ExceptionHandler(GlobalException.class)
     public ResponseEntity<ErrorResponse> handleException(GlobalException exception) {
-        ErrorResponse errorResponse = new ErrorResponse(ErrorCode.BAD_REQUEST,
-            exception.getMessage());
-        return new ResponseEntity<>(errorResponse, HttpStatus.BAD_REQUEST);
+        // todo logger 처리
+        ErrorResponse errorResponse = new ErrorResponse(exception.getErrorType(),
+            exception.getErrorType().getMessage());
+        if (exception.getErrorType().getCode() >= 500_000) {
+            return ResponseEntity.internalServerError().body(errorResponse);
+        } else if (exception.getErrorType().getCode() >= 400_000) {
+            return ResponseEntity.badRequest().body(errorResponse);
+        }
+        return ResponseEntity.badRequest().body(errorResponse);
     }
 }
