+++ src/main/java/com/example/springboard/service/CommentService.java
@@ -1,9 +1,9 @@
 package com.example.springboard.service;
 
-import com.example.springboard.common.exception.ErrorTypeWithRequest;
 import com.example.springboard.common.exception.GlobalException;
 import com.example.springboard.domain.articles.Comment;
 import com.example.springboard.repository.CommentRepository;
+import com.example.springboard.util.enums.ErrorTypeWithRequest;
 import java.util.Map;
 import org.springframework.stereotype.Service;
