+++ src/main/java/com/example/springboard/util/enums/ErrorTypeWithRequest.java
@@ -1,4 +1,4 @@
-package com.example.springboard.common.exception;
+package com.example.springboard.util.enums;
 
 public enum ErrorTypeWithRequest {
 
@@ -59,5 +59,13 @@ public enum ErrorTypeWithRequest {
     ErrorTypeWithRequest(int code, String message) {
         this.code = code;
         this.message = message;
+    }
+
+    public int getCode() {
+        return code;
+    }
+
+    public String getMessage() {
+        return message;
     }
 }
